Extract View class
